Refactor: 구조 이름 변경
